Refactor + bug fix
Move perturbation logic upstream, so it's easier to see regular output and perturbed output side-by-side
In the previous version, softmax was applied everywhere, while it should only be used for sentiment prediction. Dropout is not 0 by default, so it was applied for perturbed output, but not for the regular output.